refactor(api-query-jpa): 패키지 이동
- active 패키지로 이동
