+++ modules/infrastructure/api-query-jpa/src/main/java/com/whoz_in/api_query_jpa/device/active/ActiveDeviceJpaViewer.java
@@ -1,12 +1,8 @@
-package com.whoz_in.api_query_jpa.device;
+package com.whoz_in.api_query_jpa.device.active;
 
-import com.whoz_in.api_query_jpa.member.Member;
-import com.whoz_in.api_query_jpa.member.MemberConnectionInfoRepository;
-import com.whoz_in.api_query_jpa.member.MemberRepository;
 import com.whoz_in.main_api.query.device.application.active.view.ActiveDevice;
 import com.whoz_in.main_api.query.device.application.active.view.ActiveDeviceViewer;
 import java.util.List;
-import java.util.Objects;
 import java.util.Optional;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
